+++ Bricouli/src/main/java/com/demo/Bricouli/repositories/CommandeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.demo.Bricouli.models.Commande;
 
 @Repository
-public interface CommandeRepository extends JpaRepository<Commande, Double>{
+public interface CommandeRepository extends JpaRepository<Commande, Long>{
 
 	List<Commande> findByEtat(String s);
